backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
requests
pydantic-settings
